# BotSharp.WebHost/wwwroot/swagger.json
{
  "swagger": "2.0",
  "host": "localhost:7500",
  "basePath": "/",
  "schemes": [ "http" ],
  "info": {
    "title": "Articulate API Documentation",
    "version": "0.0.1",
    "contact": { "name": "Smart Platform Group" }
  },
...
  "paths": {
...
    "/domain/{id}": {
      "get": {
        "summary": "Find a model instance by id from the data source",
        "operationId": "getDomainId",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the domain",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "domain" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      },
      "put": {
        "summary": "Update attributes for a model instance and persist it into the data source",
        "operationId": "putDomainId",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the domain",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "schema": { "$ref": "#/definitions/Model 30" }
          }
        ],
        "tags": [ "domain" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      },
      "delete": {
        "summary": "Delete a model instance by id from the data source",
        "operationId": "deleteDomainId",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the domain",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "domain" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
    "/entity/{id}": {
      "get": {
        "summary": "Find a model instance by id from the data source",
        "operationId": "getEntityId",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the entity",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "entity" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      },
      "put": {
        "summary": "Update attributes for a model instance and persist it into the data source",
        "operationId": "putEntityId",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the entity",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "schema": { "$ref": "#/definitions/Model 32" }
          }
        ],
        "tags": [ "entity" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      },
      "delete": {
        "summary": "Delete a model instance by id from the data source",
        "operationId": "deleteEntityId",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the entity",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "entity" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
...
    "/agent/{id}/domain": {
      "get": {
        "summary": "Find list of domains linked with a model instance specified by id",
        "operationId": "getAgentIdDomain",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the agent",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "number",
            "description": "The index of the first element to return. 0 is the default start.",
            "name": "start",
            "in": "query"
          },
          {
            "type": "number",
            "description": "Number of elements to return from start. All the elements are returned by default",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "description": "String that will filter values to return only those domains with part of this filter in their names",
            "name": "filter",
            "in": "query"
          }
        ],
        "tags": [ "agent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
...
    "/agent/{id}/entity": {
      "get": {
        "summary": "Find list of entities linked with a model instance specified by id",
        "operationId": "getAgentIdEntity",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the agent",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "number",
            "description": "The index of the first element to return. 0 is the default start.",
            "name": "start",
            "in": "query"
          },
          {
            "type": "number",
            "description": "Number of elements to return from start. All the elements are returned by default",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "description": "String that will filter values to return only those entities with part of this filter in their names",
            "name": "filter",
            "in": "query"
          }
        ],
        "tags": [ "agent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
...
    "/agent/{id}/domain/{domainId}": {
      "get": {
        "summary": "Find a domain by id that belongs to the specified model instance",
        "operationId": "getAgentIdDomainDomainid",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the agent",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "number",
            "description": "Id of the domain",
            "name": "domainId",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "agent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
...
    "/agent/{id}/entity/{entityId}": {
      "get": {
        "summary": "Find an entity by id that belongs to the specified model instance",
        "operationId": "getAgentIdEntityEntityid",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the agent",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "number",
            "description": "Id of the entity",
            "name": "entityId",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "agent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
...
    "/domain": {
      "post": {
        "summary": "Create a new instance of the model and persist it into the data source",
        "operationId": "postDomain",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": { "$ref": "#/definitions/Model 2" }
          }
        ],
        "tags": [ "domain" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
    "/entity": {
      "post": {
        "summary": "Create a new instance of the model and persist it into the data source",
        "operationId": "postEntity",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": { "$ref": "#/definitions/Model 4" }
          }
        ],
        "tags": [ "entity" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
...
  },
...
}